\name{Frogs}
\alias{Frogs}
\title{Frogs Data}
\usage{data(Frogs)}
\description{
The Frogs data consist of Mel-Frequency Cepstral Coefficients (MFCCs) extracted from syllables of anuran (frog) calls. The dataset includes labels for family, genus, and species of the frogs.
}
\value{
\item{Frogs}{a data frame with 7195 rows and 22 columns}
}
\examples{
data(Frogs)
head(Frogs)
}
